flask_app/requirements.txt
tensorflow==1.15.0
keras==2.2.5
h5py
pillow
flasgger
flask
mod_wsgi